+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).  
 
+## [0.0.4] - 2017-03-22
+### Updated:
+- Production values for licenses
+- Database api fixed to handle get_queue_submissions_with_pos_for_accounts
+
 ## [0.0.3] - 2017-03-14
 ### Added:
 - New api calls: